// frontend/src/api/generated/models/salesPOResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */

export interface SalesPOResponse {
  id: string;
  customer_id: string;
  po_number: string;
  order_date: string;
  due_date?: string | null;
  status: string;
  is_locked: boolean;
  version: number;
  notes: string;
}